Coroutine for input time command

+++ Assets/Scripts/ScriptableObject/GameSettings.cs
@@ -6,8 +6,8 @@ public class GameSettings : ScriptableObject
     public int MaxPlayerNumber;
     public int TileMaxLifePoints;
 
-    [Tooltip("Temps en s")] public float CommandInputTime;
-    [Tooltip("Temps en s")] public float PlayTime;
+    [Tooltip("Temps en s")] public int CommandInputTime;
+    [Tooltip("Temps en s")] public int PlayTime;
     
     public AudioClip JumpSound;
     public AudioClip BumpSound;
